+++ src/modules/multi_ir_bgck_mod.f90
@@ -400,7 +400,7 @@ contains
     real(8) :: minpavhrr(2:3)
     real(8) :: anisot,zlamb,zcloud,scos,del,deltaphi
     integer :: ier,ijour,iloc(2:3),co2min(1),co2max(1)
-    integer :: channelIndex,ilist_sun,ilist_co2(nco2),ilist_co2_pair(nco2),ilist_he(nch_he)
+    integer :: channelIndex,ilist_co2(nco2),ilist_co2_pair(nco2),ilist_he(nch_he)
     integer :: nlv_T,id,sensorIndex, qcid, nchannels
     logical :: liasi,lairs,lcris
     type (rttov_profile), pointer :: profiles(:)